+++ package.json
@@ -2,7 +2,8 @@
 	"type": "module",
 	"scripts": {
 		"build": "npx tsc --build",
-		"start": "node --env-file=.env ./main.js"
+		"start": "node --env-file=.env ./main.js",
+		"clear": "powershell -Command \"Get-ChildItem -Path . -Recurse -Filter *.js | Where-Object { -not $_.FullName -like '*\\node_modules\\*' } | Remove-Item\""
 	},
 	"dependencies": {
 		"@types/puppeteer": "^5.4.7",
