+++ scripts/classes/Company.ts
@@ -1,4 +1,4 @@
-import Job from './Job';
+import Job from './Job.js';
 
 /**
  * Represents a company
