+++ scripts/classes/PageHandler.ts
