+++ scripts/debugUtils.ts
